refactor: 메서드 추출 및 미사용 필드 제거
