+++ src/main/java/kitchenpos/application/order/OrderService.java
@@ -1,6 +1,5 @@
 package kitchenpos.application.order;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.domain.menu.MenuRepository;
@@ -9,7 +8,6 @@ import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.order.OrderValidator;
-import kitchenpos.domain.table.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,24 +15,17 @@ import org.springframework.transaction.annotation.Transactional;
 public class OrderService {
     private final MenuRepository menuRepository;
     private final OrderRepository orderRepository;
-    private final OrderTableRepository orderTableRepository;
     private final OrderValidator orderValidator;
 
-    public OrderService(MenuRepository menuRepository, OrderRepository orderRepository,
-                        OrderTableRepository orderTableRepository, OrderValidator orderValidator) {
+    public OrderService(MenuRepository menuRepository, OrderRepository orderRepository, OrderValidator orderValidator) {
         this.menuRepository = menuRepository;
         this.orderRepository = orderRepository;
-        this.orderTableRepository = orderTableRepository;
         this.orderValidator = orderValidator;
     }
 
     @Transactional
     public OrderResponse create(final OrderRequest request) {
-        return OrderResponse.from(
-                orderRepository.save(
-                        new Order(request.getOrderTableId(), mapToOrderLineItems(request.getOrderLineItems()),
-                                orderValidator))
-        );
+        return OrderResponse.from(orderRepository.save(mapToOrder(request)));
     }
 
     public List<OrderResponse> list() {
@@ -50,15 +41,21 @@ public class OrderService {
         return OrderResponse.from(order);
     }
 
+    private Order mapToOrder(OrderRequest request) {
+        return new Order(request.getOrderTableId(), mapToOrderLineItems(request.getOrderLineItems()), orderValidator);
+    }
+
     private List<OrderLineItem> mapToOrderLineItems(List<OrderLineItemRequest> requests) {
-        var result = new ArrayList<OrderLineItem>();
-        for (OrderLineItemRequest request : requests) {
-            var menuId = request.getMenuId();
-            if (!menuRepository.existsById(menuId)) {
-                throw new IllegalArgumentException("주문 항목의 메뉴가 존재하지 않습니다. menuId = " + menuId);
-            }
-            result.add(new OrderLineItem(menuId, request.getQuantity()));
+        return requests.stream()
+                .map(this::mapToOrderLineItem)
+                .collect(Collectors.toList());
+    }
+
+    private OrderLineItem mapToOrderLineItem(OrderLineItemRequest request) {
+        var menuId = request.getMenuId();
+        if (!menuRepository.existsById(menuId)) {
+            throw new IllegalArgumentException("주문 항목의 메뉴가 존재하지 않습니다. menuId = " + menuId);
         }
-        return result;
+        return new OrderLineItem(menuId, request.getQuantity());
     }
 }
