+++ src/main/java/kitchenpos/ui/TableRestController.java
@@ -4,10 +4,10 @@ import java.net.URI;
 import java.util.List;
 import kitchenpos.application.table.ChangeEmptyRequest;
 import kitchenpos.application.table.ChangeNumberOfGuestsRequest;
+import kitchenpos.application.table.CreateTableRequest;
 import kitchenpos.application.table.OrderTableResponse;
 import kitchenpos.application.table.TableService;
 import kitchenpos.domain.table.OrderTable;
-import kitchenpos.ui.reqeust.CreateTableRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
